moved details viewmodel to appropriate package
